flathub: Trigger update-repo job when a publish job reports one
We use an in-memory cache to reuse jobs if possible, although this
is not super-critical as the job is just essentially a tail -f of
the job, so if it was duplicated that is not a huge thing.

The update job is visible both on the build page, and the publish page.