flathub: Don't check arches for extensions
Figuring out what version to use for extensions is very complicate,
involving reading the version info in the extension points in the
runtime metadata. For now lets just ignore that and force people
to manually force a specific arch set if an extension isn't available.

Also some minor cleanups and handling of custom builds (which may not
necessarily have a manifest file with the expected name.